+++ apps/marketing/src/cms/index.tsx
@@ -1,5 +1,6 @@
+import { auth } from '@clerk/nextjs/server'
 import { createTypographyResolver } from '@deepdish-cloud/resolvers/typography'
-import { configure } from '@deepdish/config'
+import { configure } from '@deepdish/ui/config'
 
 export function cms(url: string, apiKey: string) {
   configure({
@@ -7,6 +8,9 @@ export function cms(url: string, apiKey: string) {
       typography: {
         resolver: createTypographyResolver(url, apiKey),
       },
+    },
+    draft: {
+      auth: () => Boolean(auth().userId),
     },
   })
 }