+++ packages/config/src/config.ts
@@ -1,39 +1,65 @@
 import type { Result } from '@byteslice/result'
+import { getLogger } from '@logtape/logtape'
 import type { Contract } from './contract'
+import { configureLogging } from './logging'
 import type { ValueType } from './schemas'
+
+const logger = getLogger(['deepdish', 'config'])
 
 type Config = {
   contracts: {
     readonly [T in ValueType]?: Contract<T>
   }
+  draft: {
+    auth: () => boolean | Promise<boolean>
+  }
 }
 
 let config: Config
 
-function failure(message: string) {
+function configured() {
+  const message = 'DeepDish has already been configured.'
+  logger.warn(message)
+  return { failure: new Error(message) }
+}
+
+function notConfigured() {
+  const message = 'DeepDish has not been configured.'
   console.error(message)
   return { failure: new Error(message) }
 }
 
-export function configure(input: Config): Result<void> {
+export async function configure(input: Config): Promise<Result<void>> {
   if (config) {
-    return failure('Configuration has already been initialized')
+    return configured()
   }
 
   config = Object.freeze(input)
 
+  await configureLogging()
+
+  logger.info('DeepDish configured successfully.')
   return { data: undefined }
 }
 
 export function getContract<T extends ValueType>(type: T): Result<Contract<T>> {
   if (!config) {
-    return failure('Configuration has not been initialized')
+    return notConfigured()
   }
 
   const contract = config.contracts[type]
   if (!contract) {
-    return failure(`Missing contract: ${type}`)
+    logger.error('Missing {type} contract.', { type })
+    return { failure: new Error(`Missing '${type}' contract`) }
   }
 
   return { data: contract }
+}
+
+export function getDraft(): Result<Config['draft']> {
+  if (!config) {
+    return notConfigured()
+  }
+
+  return { data: config.draft }
 }