+++ packages/workbench/src/stylesheet.ts
@@ -1,0 +1,27 @@
+'use client'
+
+import { makeStyleSheet } from '@deepdish/stylesheet'
+
+export const stylesheet = makeStyleSheet('workbench', {
+  vars: {
+    // https://tailwindcss.com/docs/customizing-colors
+    slate50: '#f8fafc',
+    slate100: '#f1f5f9',
+    slate200: '#e2e8f0',
+    slate300: '#cbd5e1',
+    slate400: '#94a3b8',
+    slate500: '#64748b',
+    slate600: '#475569',
+    slate700: '#334155',
+    slate800: '#1e293b',
+    slate900: '#0f172a',
+    slate950: '#09142b',
+
+    // https://tailwindcss.com/docs/font-size
+    fontSizeXs: '0.75rem',
+    fontSizeSm: '0.875rem',
+    fontSizeBase: '1rem',
+    fontSizeLg: '1.125rem',
+    fontSizeXl: '1.25rem',
+  },
+})