+++ packages/config/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @deepdish/config
 
+## 0.2.3
+
+### Patch Changes
+
+- 6df662b: Implemented logger.
+- 6df662b: Configured logging.
+
+## 0.2.2
+
+### Patch Changes
+
+- 88a2137: Added draft mode.
+- 01c9add: Moved internal dependencies to dev dependency list.
+
+## 0.2.1
+
+### Patch Changes
+
+- 5f4b22b: Exported value type from schemas.
+- e6014aa: Utilized external package for result pattern.
+- Updated dependencies [e6014aa]
+  - @deepdish/resolvers@0.4.1
+
 ## 0.2.0
 
 ### Minor Changes